Initial code setup.
The get endpoint of the client table is already implemented as an example. When you pull the code, add a csv file with the database credentials. The format of this csv will be in the readme file. If you do not yet have a Doker container as a copy of the database in Heroku, you can use the Heroku database since the only thing that is in the project right now is only a get but as soon as possible make a container in Doker to that in the future the base in heroku will not be modified when we make the delete and updated endpoints.