ETL errors resolved.
When reading the reservations .db file I was trying to read it as reserve but it was called reservations. Additionally, the routes to access the data were wrong. To move a folder back, use ../ when ./ was being used but it did not find the data. The .db files were not being created in the modified data folder. Now they are created.