New ETL based on the previously implemented one.
It now has the function that the files that are cleaned are also mapped and entered into the tables of the database that is connected according to the credentials in the .env file. It requires a previous harcoded dictionary to be able to map the columns of the df and the columns of the tables in the base. That is why after it prints the dfs and the tables, it asks you whether or not you want to do the insertion in order to prepare the dictionary and be able to run the program a second time to be able to do the insertion successfully.