The way in which queries were made to the database when passing a file to a table was changed.
The previous way was too slow. It was impossible to pass room_unavailable from the csv to the table in the database. I was there for more than 5 minutes and it never ended. It was expected that with the local base there would be no speed problems but when doing it with the Heroku base you have to take into account how massive data transfers are done. Using COPY I was able to do the data transfer in less than 5 seconds in the Heroku database.